requirements.txt
python-uinput==1.0.1
evdev==1.9.2